refactor: CDC feature lib, move code to root module
