+++ integration-libs/cdc/components/gigya-raas/gigya-raas.component.spec.ts
@@ -1,6 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { GigyaRaasComponentData } from '@spartacus/cdc/core';
+import { CdcConfig, CdcJsService } from '@spartacus/cdc/root';
 import {
   BaseSiteService,
   CmsComponent,
@@ -9,8 +10,6 @@ import {
 } from '@spartacus/core';
 import { CmsComponentData } from '@spartacus/storefront';
 import { Observable, of } from 'rxjs';
-import { CdcConfig } from '../../root/config';
-import { CdcJsService } from '../../src/auth/facade/cdc-js.service';
 import { GigyaRaasComponent } from './gigya-raas.component';
 
 declare var window: Window;
