+++ integration-libs/cdc/root/facade/index.ts
@@ -1,0 +1,1 @@
+export * from './cdc-js.service';
