+++ integration-libs/cdc/root/index.ts
@@ -1,1 +1,3 @@
 export * from './cdc-root.module';
+export * from './config/index';
+export * from './facade/index';
