+++ integration-libs/cdc/src/auth/facade/index.ts
@@ -1,2 +1,1 @@
 export * from './cdc-auth.service';
-export * from './cdc-js.service';
